backend/pyproject.toml
[tool.poetry]
name = "codsworth"
version = "0.1.0"
description = "Helpfull robot-butler"
authors = ["Filipp Bushmanov <herooftheday82@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
uvicorn = "0.32.1"
anyio = "4.6.2.post1"
click = "8.1.7"
h11 = "0.14.0"
idna = "3.10"
sniffio = "1.1"
sqlalchemy = "^2.0.36"
databases = "^0.9.0"
asyncpg = "^0.30.0"
python-dotenv = "^1.0.1"
litestar = "^2.13.0"
pydantic = "^2.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
